fix(#237): apply modifying
