+++ apps/frontend/src/features/drawing/api/vendor.api.ts
@@ -1,19 +1,19 @@
-import { bffClient } from '@/lib/api/bff-client';
+import { apiClient } from '@/lib/api';
 import type { ApiResponse } from '@/lib/api';
-import type { DbVendorSummary, DbVendorDetail } from './vendor.types';
+import type { DbVendorSummary, DbVendorDetail } from './types';
 
 export const listVendors = async (): Promise<
   ApiResponse<DbVendorSummary[]>
 > => {
   const response =
-    await bffClient.get<ApiResponse<DbVendorSummary[]>>('/vendors');
+    await apiClient.get<ApiResponse<DbVendorSummary[]>>('/vendors');
   return response.data;
 };
 
 export const getVendor = async (
   displayName: string,
 ): Promise<ApiResponse<DbVendorDetail>> => {
-  const response = await bffClient.get<ApiResponse<DbVendorDetail>>(
+  const response = await apiClient.get<ApiResponse<DbVendorDetail>>(
     `/vendors/${displayName}`,
   );
   return response.data;
