+++ apps/frontend/src/features/drawing/hooks/index.ts
@@ -24,3 +24,4 @@ export * from './useIndexMutations';
 export * from './useConstraintMutations';
 export * from './useRelationshipMutations';
 export * from './useControlPointDrag';
+export * from './useVendors';
