+++ apps/frontend/src/features/drawing/index.ts
@@ -1,5 +1,3 @@
-export * from './api/vendor.hooks';
-export * from './api/vendor.types';
 export * from './components';
 export * from './hooks';
 export * from './types';
